requirements.txt
psutil~=7.0.0
azure-identity~=1.20.0
azure-cosmos~=4.9.0
python-dotenv~=1.0.0
